+++ apps/courses/config.json
