+++ apps/courses/views/course.swig
@@ -1,4 +1,4 @@
-{% extends '../admin.swig' %}
+{% extends '../../../src/views/admin.swig' %}
 
 {% block heading %}{{ course.name }}{% endblock %}
 {% block title %}{% block heading %}{% endblock %} – Courses – {% parent %}{% endblock %}
@@ -29,7 +29,7 @@
 				<tbody>
 					{% for user in users %}
 						<tr>
-							<td><a href="/users/{{ user._id | raw }}">{{ user.name }}</a> {% include '../users/partials/label.swig' %}</td>
+							<td><a href="/users/{{ user._id | raw }}">{{ user.name }}</a> {% include '../../users/views/partials/label.swig' %}</td>
 							<td><a href="mailto:{{ user.email }}">{{ user.email }}</a></td>
 							<td><span data-clipboard-text="{{ user.barcode }}" class="clipboard">{{ user.barcode }}</span></td>
 						</tr>
